invertor-worker/requirements.txt
aiomqtt==2.1.0
asgiref==3.8.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
paho-mqtt==2.1.0
pillow==10.3.0
psutil==5.9.8
pyparsing==3.1.2
pyserial==3.5
pysolarmanv5==3.0.2
python-dateutil==2.9.0.post0
python-pm==0.0.4
PyYAML==6.0.1
schedule==1.2.2
six==1.16.0
sqlparse==0.5.0
termtables==0.2.4
typeguard==4.3.0
typing_extensions==4.12.2
uModbus==1.0.4
